Update tweaks and returned the default affinity fitting mode to ModelA:
Added a couple of minor features, and changed the behaviour fo fit-affinity-curves to use the same baseline for all SAXS titrations, as it makes sense to assume that all apo curves should start out the same. If anything, it is the delta that should change.
This adaptable delta is coming later.